Doesn't Overwrite Pin Code
Doesn't overwrite pin code if it already exists. Won't delete pin code
if it existed before. Won't start CommuniTake 9.3.18 until 10 update tries.

+++ ResetRB/app/src/main/java/com/micronet/dsc/resetrb/modemupdater/services/CleanUpService.java
@@ -6,6 +6,9 @@ import static com.micronet.dsc.resetrb.modemupdater.ModemUpdaterService.MODEM_AP
 import static com.micronet.dsc.resetrb.modemupdater.ModemUpdaterService.UPDATE_SUCCESSFUL_ACTION;
 import static com.micronet.dsc.resetrb.modemupdater.Utils.MODEM_UPDATED_AND_CLEANED_KEY;
 import static com.micronet.dsc.resetrb.modemupdater.Utils.PIN_CODE_PATH;
+import static com.micronet.dsc.resetrb.modemupdater.Utils.PREVIOUS_PINCODE;
+import static com.micronet.dsc.resetrb.modemupdater.Utils.UPDATED_COMM_ALREADY_INSTALLED;
+import static com.micronet.dsc.resetrb.modemupdater.Utils.getBoolean;
 import static com.micronet.dsc.resetrb.modemupdater.Utils.putBoolean;
 import static com.micronet.dsc.resetrb.modemupdater.Utils.runShellCommand;
 import static com.micronet.dsc.resetrb.modemupdater.Utils.sleep;
@@ -31,27 +34,32 @@ public class CleanUpService extends IntentService {
                 // Modem Firmware Update Successful, clean up communitake and LTE Modem Updater
                 try {
                     // Remove pincode.txt
-                    File pincodeFile = new File(PIN_CODE_PATH);
-                    if(pincodeFile.exists()){
-                        boolean deleted = pincodeFile.delete();
+                    if (!getBoolean(this, PREVIOUS_PINCODE, false)) {
+                        File pincodeFile = new File(PIN_CODE_PATH);
+                        if(pincodeFile.exists()){
+                            boolean deleted = pincodeFile.delete();
 
-                        if(deleted){
-                            Log.i(TAG, "Deleted communitake pincode.");
+                            if(deleted){
+                                Log.i(TAG, "Deleted communitake pincode.");
+                            }else{
+                                Log.e(TAG, "Error deleting pincode.txt");
+                            }
                         }else{
-                            Log.e(TAG, "Error deleting pincode.txt");
+                            Log.i(TAG, "Pincode doesn't exist.");
                         }
-                    }else{
-                        Log.i(TAG, "Pincode doesn't exist.");
                     }
 
                     // Force stop communitake, clear communitake, and uninstall updater.
                     // Not sure if we need to force stop communitake or not.
-                    runShellCommand(new String[]{"am", "force-stop", COMM_APP_NAME});
-                    sleep(500);
-                    runShellCommand(new String[]{"pm", "clear", COMM_APP_NAME});
-                    sleep(500);
-                    runShellCommand(new String[]{"pm", "uninstall", COMM_APP_NAME});
-                    sleep(500);
+                    if (!getBoolean(this, UPDATED_COMM_ALREADY_INSTALLED, false)) {
+                        runShellCommand(new String[]{"am", "force-stop", COMM_APP_NAME});
+                        sleep(500);
+                        runShellCommand(new String[]{"pm", "clear", COMM_APP_NAME});
+                        sleep(500);
+                        runShellCommand(new String[]{"pm", "uninstall", COMM_APP_NAME});
+                        sleep(500);
+                    }
+
                     runShellCommand(new String[]{"pm", "uninstall", MODEM_APP_NAME});
 
                     // Update shared preferences
